ddtrace/tracer: add unit test for execution trace span annotations
Check that we create the appropriate span annotations in execution
traces. This test is dependent on parsing an execution trace, which can
change format across releases. We don't have a standard library parser.
We rely on an external implementation. To prevent this test from
breaking in the window between when a new Go version comes out and when
the parser supports any format changes, guard it with the supported Go
version. We support several Go versions, so we should have some coverage
all the time.