profiler: support using custom profiler labels in our UI
Users can use profiler labels to add annotations to CPU and goroutine
profiles. Those labels will be in the profiles we upload, but won't
currently be available as attributes for filtering frames in flame
graphs in our UI. Add an API for specifying pprof label keys which
should show up in the UI. There is a limit today to avoid worst-case
situations with massive numbers of unique label keys.