profiler: make version tag check case insensitive, deduplicate
Tags are normalized somewhere along the profile intake path. Also,
WithTags can be used to provide multiple tags with the same key, which
is behavior we should keep for compatibility in most cases. But for the
version tag case in particular, we really only want one. So, keep only
the first one we see if a tag is only provided via WithTags.