profiler: use only one service version value (#2322)
Users can set the version tag for profiles in three ways: via
WithVersion, the DD_VERSION environment variable, or via WithTags. A
service should only have one version tag at a time for profile
comparison with a "recent" version to work properly in the UI. Today,
all of these options can be used at once and the profiles for a service
can have multiple version tags.

To be consistent with the tracer, we should supply a single version and
it should be configured with the following priority order: first,
WithVersion is used. If that option is not set, then DD_VERSION will be
used. If either of those are used, then they get priority over any
version tags supplied via WithTags. That is, version tags are removed if
the version is explicitly set. Additionally, if multiple version tags are
supplied via WithTags, only the first one will be kept.