tracer: report config-change telemetry in dynamic config (#2350)
Report configuration changes applied via dynamic configuration. This required some refactoring of the current dynamicConfig implementation with an intent to keep it as generic as possible.
More logic was added to avoid unnecessary reconfigurations and unnecessary (or inaccurate) config change telemetry events in case of an attempt to reconfigure the same value.