profiler: support user-triggered trace collection
Due to the high data volume produced by the execution traces, we collect
execution traces infreuqently by default. This means our data collection
won't always align with activity the user is interested in. Add support
for a user-provided trace collection trigger. This mechanism can be
wired up to whatever mechanism is convenient to the user (such as a flag
that is set via a debug endpoint) to trigger targeted data collection.

To prevent accidentally incurring massive data transfer bills, we
rate-limit user-triggered collection on the client. The rate limiting
allows for burst of consecutive traces, while overall respecting the
configured trace period so we maintain the same average data rate. The
limit is 5 traces per 5 trace periods (5 is the "burst size" and is
subject to change). Limiting is enabled when we first see a
user-triggered trace which we otherwise would not have collected, i.e.
not randomly, and not the first profiling cycle. Limiting is disabled
again after 5 trace periods have passed. The motivations of
conditionally enabling the limiter are 1) to not penalize the user for
traces we would have already been willing to collect, and 2) to not
unduly skew the distribution of traces we normally collect randomly.

While here, we remove support for run-time reconfiguration of execution
trace collection. It is not generally safe for a user to modify
environment variables at run-time, and also probably not desirable to
allow the user to completely change the collection frequency or size
limits at runtime (will they be reliably set back to lower values?).
With the new trigger mechanism, users will have a safer and more
flexible way to target collection.