dyngo: dynamically register listeners only if they're needed
With a somewhat significant refactor, and use of Go generics, it is possible to have AppSec listeners self-install if, and only if a contrib that may emit events they listen to is loaded.

This avoid logging uncanny messages in debug mode, where we previously claimed to register all available listeners (GraphQL, gRPC, HTTP), despite some of them being nonsensical in the user's application context.

It also allows the compiler to perform better dead code elimination, or even entirely skipping packages corresponding to the unneeded listeners.