all: implement mocktracer in terms of the new API (#2409)
This PR implements the mocktracer package in terms of the new tracing
APIs. The goal here is to have as few changes as possible to the
mocktracer API and therefore to the contrib tests.

This is not yet complete, but it is working for the gin-gonic contrib
package. This is a good example of what I'm going for and the (hopefully)
minimal changes that will be required for the other contrib packages.

These changes need to go along with DataDog/system-tests#1935 which
update the parametric test to use the new API. For now, that branch has
been hard-coded in the CI for this PR, but that will be changed before
merging.