maitenance: new GitHub Triage Badger (#2451)
Adds a new GitHub workflow which automatically adds a `needs-triage` label to issues that are opened or re-opened, which is to be removed whenever triage is performed.

Every weekday (at 8am UTC, per the current configuration), the workflow lists up to 10 issues that have the `needs-triage` label and are more than 3 days old, and sends messages to Slack via a configrued WebHook to pester maintainers to triage them. The Slack message includes a link to the issue as well as the issue's ID, title, body, and creation timestamp.