appsec: register ASM_TRUSTED_IPS capability to RC
ASM_TRUSTED_IPS is the Remote Config capability that signifies this
library is compatible with trusted IPs; meaning it uses
`libddwaf v1.12.0` or later.

This is expected to produce additional data under the `ASM_DATA` RC
product, which contains a set of trusted IPs. The merge strategy for
this data is identical to that of other data currently received from RC,
meaning if the same entry is provided multiple times, the library will
retain the one with the latest expiration timestamp.