ddtrace/tracer: clear global headers in remote config tests
Clear the global headers in the remote config tests which modify them.
Otherwise, tests like TestWithHeaderTags, which are sensitive to the
current global headers, can fail if they're run after the remote config
tests.

This might not be the ideal solution to this problem in general, but it
at least covers the other existing case of a test which reads these
values. Once we start shuffling unit tests we'll ideally catch these
kinds of issues as they're introduced, and we can come up with a more
robust way to prevent them if it keeps happening.