profiler: support using custom profiler labels in our UI (#2282)
Users can use profiler labels to add annotations to CPU and goroutine
profiles. This helps them contextualize profiling information.
Those labels will be in the profiles we upload, but won't currently be
available as attributes for filtering flame graphs in our UI. We have support
for that now in our backend & UI, so we just need a way to make it
available to users. Add an API for specifying profiler label keys which should
show up in the UI. There is a limit today to avoid worst-case situations with
massive numbers of labels.