chore: make the smoke-test workflow usable as go-libddwaf integ test
In order to prevent reoccurrences of issues where non-go files are
located in places incompatible with `go mod vendor`, which the smoke
test accurately detects, we are going to run the smoke tests as part of
the CI suite in go-libddwaf itself. This requires being able to pin the
go-libddwaf version to the commit being CI'd.