chore: fix smoke test workflow_call situation
The default repository being checked out by `actions/checkout` is the
one designated by `${{ github.repository }}`, which is the repository
where the original event (at the root of the workflow call tree)
started.

We however must check out the source to `dd-trace-go` and not that of
the caller workflow. This requires hard-coding the repository name here.