Extract and set _dd.parent_id from the tracestate
The goal here is to get the _dd.parent_id tag to
be set on a span created from this context.

For example, if we create a span that is a child
of this context (a context from W3C), we want
to encode the value of `p` from that context
into that span's Meta.