Changes to dynamic instrumentation remoteconfig callback logic
This makes changes to the logic of how the remote config client handles
updates to dynamic instrumentation configurations. Since DI is hooking
the callback via bpf to extract the config, we need to be able to get
the current config from memory in the case that that DI has started
after the tracer received the config update.

Signed-off-by: grantseltzer <grantseltzer@gmail.com>