internal: correct agent resolution comment, clarify order in code
The comment describing the agent URL resolution did not specify what
happens if only one of DD_AGENT_HOST or DD_TRACE_AGENT_PORT are defined.
Do so, according to the internal RFC for agent URL resolution. Rearrange
the code so that it more clearly matches the priority order. Before this
commit, the check for whether the host or port are defined was not
clearly visible, hidden away in the check for the existence of the
default UDS path.

Add some more tests to check the host/port behavior. Also, the assert
package functions expect the "want" before the "got", so fix that up in
the test cases.