contrib/database/sql: remove broken execution trace tests
The execution tracer tests for contrib/database/sql use a parser which
does not support Go 1.21 or newer traces. The tests should be ported to
use golang.org/x/exp/trace and moved to internal/exectracetest, but for
now they are effectively broken since we no longer run Go 1.20 in CI.
Remove them to fix CI on main, until we port them to the new parser.