internal/exectracetest: port database/sql test to the new parser
Port the contrib/database/sql execution tracer test to use the new
golang.org/x/exp/trace parser. The test is slightly cleaned up, but for
the most part covers the same logic as its predecessor (see #2060).  The
primary difference is that, rather than hooking into a mock database/sql
driver, we use a real one (sqlite3). To simulate a slow operation, the
test registers a sleep function and also sleeps on connection. The
purpose of this is to ensure, where we can, the tracer tasks cover the
whole operation and not just the end of it. We no longer hook into
_every_ operation, though, so we drop some time assertions.