all: upgrade minimum supported Go release to 1.22 (#2850)
With Go 1.23 released on 2024-08-13, we drop support for Go 1.21. This
means Go 1.22 is now our minimum supported Go release. Update our CI
files to reflect our current supported Go versions, and remove pre-Go
1.22 workarounds.

In addition to the CI bumps, this PR changes/fixes a few things related to the upgrade:

- Drops the old `rand` from the tracer, now that we can depend on `math/rand/v2`.
- Uses `slices` in the appsec package.
- The `TestWrapHandlerWithResourceNameNoRace` got a new data race. Specifically,
  by bumping our `go.mod` to 1.22, we get the new 1.22 ServeMux behavior internally.
  The 1.22 ServeMux modifies the Request, introducing a race with all the goroutines
  accessing/modifying the request concurrently. Fix it by giving each goroutine its own
  request. I verified that the test still catches the intended bug by reverting the fix from
  #2204 and verifying that the test fails.