internal/apps: Use latest go version (1.23)
The internal apps are usually meant to demonstrate new features or to
produce screenshots for blog posts. In both cases it is better to be
using the latest version of Go as it might contain new features that are
needed, or simply to make the screenshots show an up-to-date Go version.