internal/apps: move setup-smoke-test
Move setup-smoke-test from internal/apps/setup-smoke-test to
internal/setup-smoke tests. This allows the other internal/apps to use
the latest version of go. We want this because these apps are often used
to test the latest new features which may depend on new go version. They
are also used for screenshots where we don't want to show an old go
version.

setup-smoke-test is serving a different purpose and needs to be tested
against the minimum version of Go that we support. So move it to a
separate location and give it its own go.mod file.