internal/appsec/emitter/waf: reduce rate limiting log spam
The ContextOperation.AddEvent method logs a warning if the
ContextOperation's rate limit is hit. If there are a burst of calls to
AddEvent, this means there will be a corresponding burst of logs. This
happens for example when running BenchmarkGraphQL on the benchmarking
platform. Only the first log is really useful.