Add logic to access strings used by remote config bpf probe
This adds logic to the remote-config callback logic, which is used by
the Go dynamic instrumentation product, to access the probe definition
strings before they are passed to the hooked callback. This is to
ensure that the strings are in memory. If they are not in memory, or
not in the same thread page table, page faults can occur which causes
bpf_probe_read() calls to fail. This is in a critical path for Go DI
so this change will help mitigate that as much as possible.

Signed-off-by: grantseltzer <grantseltzer@gmail.com>