internal/orchestrion/_integration: go-redis tests now wait for successful Ping
There are frequent flakes on these tests caused by networking errors
that typically arise from the server not being completely ready before
the tests start. To alleviate this issue, we now wait for a successful
ping during the setup phase to make sure we have a valid client before
proceeding to tests.