internal/orchestrion/_integration: enable vm.overcommit_memory
This is necessary to safely operate a REDIS server, as it will easily trigger the OOM
reaper under tight memory conditions (often the case on CI machines...), but these
are artificial because the process maps for fork children is copy-on-write, and should
be much leaner than they appear to be. Hopefully this reduces/removes flakes.