fix(appsec/gqlgen): recovered from an unexpected panic from an event listener
The GraphQL subscription propagates a `nil` span to avoid creating an indefinitely long-running span. This results in a nil-panic later on as tags are being set on this.

To address this, override the `TagSetter` with a `NoopTagSetter` to avoid dealing with nil checks everywhere.

Added a subscriptions test on the gqlgen contrib that I used to validate the panic no longer triggers.