ddtrace/tracer: more accurate retry logging
When the writer fails to send a payload to the agent, it logs an error message
that includes the text "will retry", but technically this is only true
if the tracer was started with the `WithSendRetries` option (otherwise,
the config value is 0 and no retries will be attempted).

Would it be more accurate/less confusing to do something like this where
the writer just logs "attempt N of M"?