profiler: clear old completed list before re-using it (#3304)
Make sure the completed profiles list is cleared before re-using it. The goal is to release the pointers to old profiles to the garbage collector instead of keeping them reachable during the new profiling loop.

This is especially important when execution-trace is enabled because this profile type is not collected at every run of the loop, meaning that you can have old profiles remaining in the underlying array for multiple profiling cycles.