appsec: set HTTP header span tags even if WAF has no interest in them
Some products rely on the presence of the HTTP header span tags to be
collected. Currently, these are only collected if the WAF is interested
in any of the HTTP server addresses; but if that is not the case, these
addresses are not set. This is cause System Tests for the ATO SDK to
fail as the feature assumes `http.client_ip` is present.

This registers a "basic HTTP" feature that does the HTTP header
extraction and tag setting when the full-on security implementation is
not enabled; so that the headers are always collected.