contrib/redis/rueidis: add Mode method support (v1.0.56)
Add a Mode method on the client struct so that it implements the
updated rueidis.Client interface. This fixes a compile error that
occurred when using the latest version of rueidis. See
https://github.com/DataDog/dd-trace-go/discussions/3347.

Also update go.mod and go.sum to bump github.com/redis/rueidis to
v1.0.56, along with related dependency upgrades.