refactor(profiler): centralize compression logic
For historical reasons, the profiler has used varying levels of
compression for different profile types (no compression, gzip-1 and
gzip-6). This has been identified as target for cost optimization,
see PROF-11815.

Refactor the compression related code in order to pave the path for
introducing a new compression strategy, which will most likely be based
on zstd.

This patch should not have any observable impact. Profiles are expected
to be uploaded with exactly the same compression settings as before, and
the performance is expected to stay exactly the same as well.

Introduce a dedicated compressor for each profile type that handles
converting the profile data produced by it from the existing compression
(gzip-1 or no compression) to the desired output compression.

Update delta profiles and expGoroutineWaitProfile to perform their
gzip-6 compression in the compressor for their profile type, rather than
doing it themselves.

A few TestRunProfile tests had to be modified because they were not
explicitly declaring the profile types they were testing when
initializing the profiler, which caused panics due to their compressors
being uninitialized.