profiler: support zstd compression (#3555)
Support zstd compression of all profile types, including re-compression
from gzip to zstd for the profile types which are already compressed by
the runtime. The compression is configured by the environment variable
DD_PROFILING_DEBUG_COMPRESSION_SETTINGS, which is of the form
algorithm-level. The level is optional, and the supported algorithms are
gzip and zstd. If provided, the selected algorithm applies to all uploaded
profiles. This is intended for internal testing. Based on the results, we will
probably update the default algorithm and stop supporting this configuration.

Co-authored-by: Felix Geisendörfer <felix.geisendoerfer@datadoghq.com>