profiler: stop using runProfile in tests
In general, we should prefer testing the profiler with as close to the
API customers will use as possible. For one, it gives us more confidence
that customers will get the data they expect. Also, it's been a pain to
work on the profiler when internals-only changes break tests that happen
to depend on the internals of the profiler, such as TestRunProfile. This
change starts getting rid of uses of runProfile in particular.

This change gets rid of the CPU and goroutine sub-tests in
TestRunProfile since they really don't assert anything meaningful (other
tests already check that those profiles are _present_, and mocking their
contents doesn't tell us anything on top of that). For the goroutine
wait profile sub-tests, we can at least remove the dependency on run
profile. This change moves those sub-tests to their own test functions,
TestGoroutineWait and TestGoroutinesWaitLimit. For TestGoroutineWait, we
are basically testing converting a debug goroutine profile to a pprof,
so the tests opts to use goroutineDebug2ToPprof directly rather than
hook into the profiler. For TestGoroutinesWaitLimit, we use
startTestProfiler, which is more end-to-end, and make assertions about
the profiles we would upload and the logs we would emit instead of
hooking in to the profiler.