ddtrace/tracer: cap execution tracer task name length
We use the span resource as the name for execution tracer task
annotations. The resource can be an arbitrarily large string, like an
SQL query, and the execution tracer can fail to allocate space for it in
its string table and crash. The task name isn't really used for anything
right now other than debugging. Cap the task name at some arbitrary
length to mitigate this.