profiler: fix race in refreshing execution trace config
We have a race in the profiler between reading the execution trace
configuration during upload and refreshing it in the profiling loop.
This shows up with most of our tests when run under the race detector.
Re-reading the trace configuration after starting the profiler seems
like a bad idea in hindsight: for that to be useful, the process would
need to update its environment variables at run time, and reading and
writing env vars concurrently is a no-no: https://www.evanjones.ca/setenv-is-not-thread-safe.html
Fix the race by just not doing that.