ci: actually fail if core tests fail
The ci-test-core.sh script included `set +e`, which turns _off_ failure
if any pipeline in the script fails. That's not quite what we want.
We want to run both of the test commands so that we can see all of the
failures for a PR, but we still want to exit with a non-zero status if
one of them fails. The ci-test-contrib.sh does this right, so copy that.