profiler: skip flaky TestDebugCompressionEnv test
The flakes we've seen are errors of the form

	profiler_test.go:363: bad multipart form: unexpected EOF

I don't think this error case is necesarily specific to this test (maybe
there's a "race" between stopping the test profiler and mock backend
server) but this test is where we're seeing it. Skip it for now to keep
CI green while we root-cause the failures.

Updates #3681